Refactored Dewey methods to separate class.
